model/requirements.txt
numpy>=1.19.2
tensorflow>=2.4.0
keras>=2.4.3
Pillow>=8.0.1
scikit-learn>=0.23.2
opencv-python>=4.4.0 